requirements.txt
langchain
langchain_community
langchain_core
langchain_experimental
langchain_ibm
neo4j
json_repair
langchain
langchain_community
langchain_core
langchain_experimental
langchain_ibm
neo4j
json_repair
langchain_openai
pypdf
fsspec
langchain_neo4j
langchain_cohere
neo4j_graphrag
cohere
langchain_cohere
neo4j-graphrag[sentence-transformers]
ibm_watsonx
ibm_cloud_sdk_core
ibm-watson
ollama
tavily-python
igraph
leidenalg
numpy
flask 
flask_session
datasets
playwright
langchain_huggingface
langchain_elasticsearch
elasticsearch
community
gdsfactory
spacy
py2neo
nltx
PyPDF2
bs4
langchain_nvidia_ai_endpoints